init: org-agenda: bind agenda search
When using multiple custom agenda commands `org-agenda-list' will retain
the settings of the last custom command during any command that rebuilds
the agenda until you select a new command or call `org-agenda'.

Not sure if this is a bug, or intentional, but the workaround is to
partially apply `org-agenda' instead of calling `org-agenda-list'.