init: tab-bar: Use default tab-bar-show value
Ideally, this would work as described. With a value of 1, we expect the
tab-bar to show only when there is more than one tab available. However,
when adding a new tab, this isn't always the case and I have to manually
reload tab-bar-mode to get it to work. I can live without it until I
look into this later.