Account for multiple packages hosted in single repo
We need to block on all packages which rely on a repo that is already being downloaded.

Otherwise they will throw an error when attempting to clone their dependencies
because their source files may not be available yet.

We also want to skip checking out any ref because then we could have multiple
packages in conflict about what the correct ref is.
Need to think more on this behavior and how we want to warn/document it.