Avoid unnecessary rebuilds
This will need to be fined tuned.
Right now we assume the build directory is in good condition so long as it is on disk.
This may be overly optimistic.

We also do a basic check to see if the cached recipe has changed.
Will have to measure start up performance and go from there.