parcel-run-next-build-step: Prevent duplicate status changes
We don't want to check the status of any dependents which are already finished.
This is unnecessary work and can actually break transitive dependency builds.

If an order has failed, don't update its status to "finished".
Failure bubbles up to dependents, but this was making the dependency look like
it finished successfully.