parcel--queue-dependencies: Prevent race conditions
The order needs to have its full dependency list before any of its dependencies
are started. Otherwise a dependency which finishes before other dependencies
are added to the dependency list will cause the dependent order to unblock
itself early (it thinks all its dependencies have finished).

I've also changed the order of the build steps when a queuing a built package.
It makes more sense to ensure all dependencies are activated prior to activating
dependent orders.

TODO: This still needs to be fixed in parcel--clone-dependencies.
It's less much likely to occur because cloning and the rest of the build process
tends to takes longer than queuing the rest of the dependencies, but it's
still possible (perhaps when cloning from local network or disk?) and preventable.