parcel--process-filter: print tail of raw result
The current output chunk may be an incomplete line.
Note there's still the chance that the tail of the raw result is incomplete,
but this improves the situation a bit.
We'll have to do something more involved here (keeping track of the raw input
and logging whole lines as they occur), but I have to think about the design and
the speed implications.