Add busy process detection.
This allows us to detect when a process is still live, but possibly awaiting
input. I still have to do more testing to see if this is a suitable solution,
but it's the simplest general solution (doesn't revolve around analyzing the
process output) I can think of at the moment.