Allow multiple queues during init
Allows user to install/configure packages which will alter parcel prior to
processing the rest of the queue. This enables menus to be distributed as
separate packages and packages like keychain-environment (which caches ssh
credentials) to be installed first.

Needs work. I'm not sure if I like the current design of keeping a single queue
list but tagging orders with IDs. We could instead actually create multiple
queue lists as their own objects, which would allow us to cache associated
information on them (whether or not the queue was created during init). Have to
experiment with that approach to see if it's conceptually simpler/doesn't slow
down init too much.