parcel-order: prevent recipe lookup failure masking
Without a recipe, the repo-dir is guaranteed to fail.
Because this is checked second, our order info will show that failure.
We want to display the initial failure.
Another option would be to create a backtrace-like list of errors,
but that will probably exceed a single line in the status buffer.