elpaca--checkout-ref, elpaca--clone: account for :ref + :depth
The :depth recipe keyword may make a :ref unreachable.
The simplest solution is to ignore :depth when :ref is used.
An alternative would be to git fetch --deepen N until :ref is reachable or the
repo's history is exhausted.
Apparently in newer versions of git on can fetch with --depth N + a commit hash,
but this has not been widely adopted at this time.