README: improve bootstrap snippet
We want the user to be able to specify a recipe for Elpaca itself, but we need
Elpaca to interpret that recipe.

This version of the bootstrap snippet allows one to specify the recipe and
manually parses a subset of the recipe (the :ref and :url keywords).
That should be enough for most users to clone their desired version of Elpaca.

I've improved error handling/warning during the bootstrapping phase as well.