elpaca--process-busy-interval: increase to one minute
Previous value was too aggressive.

:pre/:post-build commands are currently run synchronously in a subprocess.
That means the output (stdout or stderr) isn't dumped until the process has
finished running. A better strategy will be to run these in a subprocess from
the subprocess.
That way we can have a filter function which messages process output back to the
main process.