log :pre/:post-build steps asynchronously
Add elpaca-process-poll to call a program asynchronously in an Emacs subprocess.
The subprocess then runs any commands which require subprocesses asynchronously
as well. However, it blocks via accept-process-output and messages changes to
the subprocess buffer during after-change-functions. The upshot is that messages
are picked up by the parent Emacs process as they come in rather than waiting
for the process to execute synchronously and dumping the log all at once.

Remove unused elpaca-with-async-process macro.