elpaca-recipe: add optional ITEMS parameter; handle nil item case
ITEMS passed to elpaca-menu-item internally.
Should allow for more flexible uses.

One thing to keep in mind is that, when used in a lisp program,
the recipe is being recomputed. So if the ITEMS include recipes which have
already been computed, e.g. elpaca--custom-candidates, elpaca-order-functions
and elpaca-recipe-functions will be merged with the computed recipe.
In this case, those hooks should be let-bound to nil to avoid overwriting the
user's declared props:

(let (elpaca-order-functions elpaca-recipe-functions)
  (elpaca-recipe nil (elpaca--custom-candidates)))

(elpaca-recipe nil) in a lisp program prompts for the recipe as promised by the documentation
now, too.