elpaca<-create: ensure mono-repo registered regardless of queue
The previous logic did not return the mono-repo unless the elpacas shared a
queue. We don't want to block in that case, but we do want to register the
mono-repo. Otherwise, subsequent sibling elpacas will fail because their
build-steps will include elpaca--clone despite the repository already being on
disk. Attempting a clone used to not cause an order to fail, but it does now
since I removed locale-specific output parsing (it's hard to tell why git failed
sometimes because it does not always return a different exit code for each type
of failure).