Allow linking builds by copy
Add elpaca-no-symlink-mode.
When enabled, recipe :files are "linked" by copying.
When the mode is active and a file is found which has a corresponding source
file in the package repo, the source file is visited.
This effectively simulates the symlink behavior on systems, such as Windows,
which make it difficult for the user to enable symlinkinks.

See: https://github.com/progfolio/elpaca/issues/72