elpaca--dependencies: set default-directory in recycled buffer
If the main elisp file is not declared, the fallback methods which
guess at the main file via `directory-files' do not return full file paths.
Prior to recycling the buffer in which the dependencies are parsed, this did not matter
because it was guaranteed to have the same default-directory.
A recycled buffer will retain its buffer-local value for default-directory, which can
cause insert-file-contents-literally to barf when the file does not exist.

See: https://github.com/progfolio/elpaca/issues/178