elpaca-generate-autoloads: fail gracefully when build dir empty
This can occur when the recipe's :files conflict with the structure of the repository.
e.g. Attempting to use an inherited :files spec which assumes a different repo layout.
Previously, loaddefs-generate would be passed a nil DIR arg, which would signal
an vague error. Note this change does not affect Emacs < version 29.