Be more lenient with core date warning for development versions
Consider Major and Minor version numbers when matching `elpaca--emacs-releases`.
Give a version specific warning if a development build is detected, but the
previous stable release's core date is assigned. This hopefully strikes a
balance between being too "loud" about this type of warning and providing a
practical fallback value for elpaca-core-date on development Emacs versions.
Related: #235

Co-authored-by: Psionik K <73710933+psionic-k@users.noreply.github.com>