elpaca--queue-dependencies: add version check when dep not built
This covers the case where the dependent is built, but one of its dependencies
is not (e.g. the user deleted the dependency in a previous session).
Previous behavior would lead to situations where dependency may still not
satisfy the minimum requirement, but dependent was not checking it's dependency versions.