elpaca-use-package: refactor to use use-package's :ensure keyword
This is less hacky than introducing our own :elpaca use-package keyword.
It also obviates the need for a separate variable to duplicate the behavior of use-package-always-ensure.
There are plenty of packages out there which should also work fine with :ensure
t.
The intent here is to make using/switching to Elpaca easier.

To migrate from the :elpaca use-package keyword, replace :elpaca with :ensure,
and elpaca-use-package-by-default with use-package-always-ensure in your
configuration files.