elpaca-process-queues: ensure elpaca-after-init-time/hook set/run
If `elpaca-wait` is called at the end of an init file, there is no
way for it to tell if it is the last queue to process, so it does not
set elpaca-after-init-time, or run elpaca-after-init-hook.

elpaca-process-queues will take care of setting the time and running the hook if
it detects that there are no incomplete queues and they have not been run/set already.

See: https://github.com/progfolio/elpaca/issues/260