speed up loading of autoloads
"Note that .el files are supposed to use UTF-8 since Emacs-25.
[...]
Once all files use UTF-8 we can get rid of `load-with-code-conversion`
which is the indirection which slows down loading of `.el` files
compared to `.elc` files.  In the mean time you can

    (let ((load-source-file-function nil)) ...)

around your call to `load` to speed it up.
" ~ Stefan Monnier
https://yhetil.org/emacs-devel/jwvjz7rh1xz.fsf-monnier+emacs@gnu.org/